+++ example/simple_bloom_filter_example.dart
@@ -1,7 +1,7 @@
 import 'package:simple_bloom_filter/simple_bloom_filter.dart';
 
 void main() {
-  var bloom = simple_bloom_filter(10000, 3);
+  final bloom = SimpleBloomFilter(10000);
 
   bloom.add('hello world');
   bloom.add(1234567890.toString());
@@ -9,26 +9,22 @@ void main() {
 
   if (bloom.check('hello world') == true) {
     print('"hello world" PROBABLY exists...');
-  }
-  else {
+  } else {
     print('"hello world" DEFINITELY does not exist...');
   }
   if (bloom.check(1234567890.toString()) == true) {
     print('"1234567890" PROBABLY exists...');
-  }
-  else {
+  } else {
     print('"1234567890" DEFINITELY does not exist...');
   }
   if (bloom.check('{"first-name": "Dart", "last-name": "lang"}') == true) {
     print('"{"first-name": "Dart", "last-name": "lang"}" PROBABLY exists...');
-  }
-  else {
+  } else {
     print('"{"first-name": "Dart", "last-name": "lang"}" DEFINITELY does not exist...');
   }
   if (bloom.check('{"first-name": "Go", "last-name": "lang"}') == true) {
     print('"{"first-name": "Go", "last-name": "lang"}" PROBABLY exists...');
-  }
-  else {
+  } else {
     print('"{"first-name": "Go", "last-name": "lang"}" DEFINITELY does not exist...');
   }
 }