+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.0.0
+
+- Nullsafety
+- Renamed to SimpleBloomFilter
+- Added addAll and addInt/checkInt functions
+- Removed hashList and use hashcode instead to improve performance drastically
+
 ## 0.1.3
 
 - Fix name in README